+++ tests/ui/no_comma_after_factory_name.rs
@@ -1,0 +1,10 @@
+use dst_factory::make_dst_factory;
+
+#[make_dst_factory(create pub)]
+struct NoCommaAfterFactoryName {
+    id: i32,
+    data: str,
+}
+
+fn main() {
+}