+++ tests/ui/no_comma_after_iterator.rs
@@ -1,0 +1,10 @@
+use dst_factory::make_dst_factory;
+
+#[make_dst_factory(iterator = MyIter pub)]
+struct NoCommaAfterIterator {
+    id: i32,
+    elements: [u8],
+}
+
+fn main() {
+}