+++ tests/ui/no_comma_after_no_std.rs
@@ -1,0 +1,10 @@
+use dst_factory::make_dst_factory;
+
+#[make_dst_factory(create, no_std deserialize)]
+struct NoCommaAfterNoStd {
+    id: i32,
+    data: str,
+}
+
+fn main() {
+}